LangGraph_Course/requirements.txt
langgraph
langchain_community
langchain
python-dotenv
ipykernel
deepinfra